remove unused & update readme
