+++ keyboards/loga/yaksa75al/config.h
@@ -1,5 +1,4 @@
 // Copyright 2024 yangzheng20003 (@yangzheng20003)
-// Copyright 2024 keRLos (@keRLos)
 // SPDX-License-Identifier: GPL-2.0-or-later
 
 #pragma once
@@ -99,3 +98,13 @@
 #define EECONFIG_USER_DATA_SIZE    (EECONFIG_RGBREC_USE_SIZE + EECONFIG_CONFINFO_USE_SIZE)
 #define RGBREC_EECONFIG_ADDR       (uint8_t *)(EECONFIG_USER_DATABLOCK)
 #define CONFINFO_EECONFIG_ADDR     (uint32_t *)((uint32_t)RGBREC_EECONFIG_ADDR + (uint32_t)EECONFIG_RGBREC_USE_SIZE)
+
+#define RGB_MATRIX_BAT_VAL      150
+
+#define IM_BAT_REQ_LEVEL1_VAL   50
+#define IM_BAT_REQ_LEVEL1_COLOR 0x00, RGB_MATRIX_BAT_VAL, 0x00
+
+#define IM_BAT_REQ_LEVEL2_VAL   30
+#define IM_BAT_REQ_LEVEL2_COLOR 0x00, RGB_MATRIX_BAT_VAL, 0x00
+
+#define IM_BAT_REQ_LEVEL3_COLOR RGB_MATRIX_BAT_VAL, 0x00, 0x00
