+++ keyboards/loga/yaksa75al/keymaps/default/keymap.c
@@ -27,7 +27,7 @@ const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
         EE_CLR, KC_MSEL, KC_VOLD, KC_VOLU, KC_MUTE, KC_MSTP, KC_MPRV, KC_MPLY, KC_MNXT, KC_MAIL, KC_WWW_HOME, KC_CALC, KC_WWW_SEARCH, KC_INS, KC_MUTE,
         _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, RGB_TOG, KC_END,
         _______, KC_BT1, KC_BT2, KC_BT3, KC_2G4, _______, _______, _______, _______, _______, _______, _______, RGB_HUI, RGB_MOD, KC_PSCR,
-        _______, _______,  TO(MAC_BASE), _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, KC_PAUS,
+        _______, _______,  TO(MAC_BASE), _______, _______, _______, _______, _______, _______, _______, _______, _______, CHK_BAT, KC_PAUS,
         _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,  MO(ADVANCED_MODE), RGB_VAI,
         _______, 0x700b, _______, HS_BATQ, _______, _______, _______, RGB_SPD, RGB_VAD, RGB_SPI
     ),
@@ -45,7 +45,7 @@ const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
        EE_CLR, KC_BRID, KC_BRIU, _______, _______, RGB_VAD, RGB_VAI, KC_MPRV, KC_MPLY, KC_MNXT, KC_MUTE, KC_VOLD, KC_VOLU, KC_INS, KC_MUTE,
        _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, RGB_TOG, KC_END,
        _______, KC_BT1, KC_BT2, KC_BT3, KC_2G4, _______, _______, _______, _______, _______, _______, _______, RGB_HUI, RGB_MOD, KC_PSCR,
-       _______, TO(WIN_BASE), _______,  _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, KC_PAUS,
+       _______, TO(WIN_BASE), _______,  _______, _______, _______, _______, _______, _______, _______, _______, _______, CHK_BAT, KC_PAUS,
        _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,  MO(ADVANCED_MODE), RGB_VAI,
        _______, _______, _______, HS_BATQ, _______, _______, _______, RGB_SPD, RGB_VAD, RGB_SPI
     ),
