+++ keyboards/loga/yaksa75al/post_rules.mk
@@ -1,5 +1,2 @@
-RGB_MATRIX_CUSTOM_USER = yes
-RGBLIGHT_DRIVER = custom
-include keyboards/loga/yaksa75al/rgb_record/rgb_record.mk
 include keyboards/loga/yaksa75al/wls/wls.mk
 include keyboards/loga/yaksa75al/wireless/wireless.mk
