--- keyboards/loga/yaksa75al/rgb_matrix_user.inc
@@ -1,18 +1,0 @@
-// Copyright 2024 yangzheng20003 (@yangzheng20003)
-// SPDX-License-Identifier: GPL-2.0-or-later
-
-#ifdef ENABLE_RGB_MATRIX_RGBR_PLAY
-RGB_MATRIX_EFFECT(RGBR_PLAY)
-#    ifdef RGB_MATRIX_CUSTOM_EFFECT_IMPLS
-
-bool RGBR_PLAY(effect_params_t *params) {
-    RGB_MATRIX_USE_LIMITS(led_min, led_max);
-
-    extern void rgbrec_play(uint8_t led_min, uint8_t led_max);
-    rgbrec_play(led_min, led_max);
-
-    return rgb_matrix_check_finished_leds(led_max);
-}
-
-#    endif // RGB_MATRIX_CUSTOM_EFFECT_IMPLS
-#endif
