--- keyboards/loga/yaksa75al/rgb_record/rgb_rgblight.h
@@ -1,3 +1,0 @@
-#include "rgb_matrix.h"
-
-bool rgb_matrix_indicators_advanced_rgblight(uint8_t led_min, uint8_t led_max);
